test(promote): split the functional harness and cleanup suite; document image retention
The functional test crossed the 800-LOC module ceiling with the retention
case, so the shared harness (fake docker/curl, scratch git repo, signed
envelope, runPromote) moves to promote-script-functional.test-support.ts
and the two cleanup-sweep tests move to promote-script-cleanup.test.ts.
No test body changed; 66 tests pass across the four promote-script files.

The self-upgrade runbook gains a cleanup checkpoint naming PROMOTE_IMAGE_KEEP
and the expected post-run tag count.

Backlog item: BI-1172E86A
Workroom: WC-7B575D5B
Convergence-Impact-Decision: self-upgrade-step — promote.sh step=cleanup now untags superseded version images on every existing install at its next self-upgrade; no operator action
