+++ apps/web/lib/self-upgrade/promote-script-functional.test.ts
@@ -1,285 +1,15 @@
 import { describe, it, expect } from "vitest";
-import { execFileSync, spawnSync } from "node:child_process";
-import { createHash, createHmac } from "node:crypto";
-import { copyFileSync, existsSync, mkdtempSync, writeFileSync, chmodSync, mkdirSync, rmSync, readFileSync, realpathSync, readdirSync, statSync } from "node:fs";
-import { tmpdir } from "node:os";
+import { readFileSync, rmSync, writeFileSync, mkdirSync, readdirSync, statSync, copyFileSync, chmodSync, existsSync } from "node:fs";
 import { join, resolve } from "node:path";
+import { createHash } from "node:crypto";
 
 // High-fidelity proof of the real promote.sh against a real Git repo, faking
 // only Docker and HTTP. The fake portal reports the stamped DPF_VERSION, so
 // sha-verify exercises the production build-arg -> image -> endpoint contract.
+// Harness lives in promote-script-functional.test-support.ts; the cleanup
+// sweep suite lives in promote-script-cleanup.test.ts.
 
-const SCRIPT = resolve(__dirname, "../../../../scripts/promote.sh");
-const REPO_ROOT = resolve(__dirname, "../../../..");
-const MIGRATOR = join(REPO_ROOT, "scripts", "installer", "migrate-install-state.mjs");
-const CATALOG = join(REPO_ROOT, "scripts", "capability-service-catalog.generated.json");
-const gitBash = join(process.env.ProgramFiles ?? "C:\\Program Files", "Git", "bin", "bash.exe");
-const BASH_COMMAND = process.platform === "win32" && existsSync(gitBash) ? gitBash : "bash";
-const BASH_OK = spawnSync(BASH_COMMAND, ["--version"], { encoding: "utf8" }).status === 0;
-const GIT_OK = spawnSync("git", ["--version"], { encoding: "utf8" }).status === 0;
-const PROMOTE_TEST_TIMEOUT_MS = 30_000;
-
-let bashDrivePrefix: string | undefined;
-
-function toBashPath(value: string): string {
-  if (process.platform !== "win32") return value;
-  const normalized = value.replace(/\\/g, "/");
-  const drivePath = /^([A-Za-z]):\/(.*)$/.exec(normalized);
-  if (!drivePath) return normalized;
-
-  const drive = drivePath[1].toLowerCase();
-  const rest = drivePath[2];
-  if (bashDrivePrefix === undefined) {
-    const cwdDrive = /^([A-Za-z]):/.exec(process.cwd())?.[1]?.toLowerCase();
-    const pwd = spawnSync(BASH_COMMAND, ["-lc", "pwd"], { encoding: "utf8" }).stdout.trim();
-    bashDrivePrefix = cwdDrive && pwd.startsWith(`/mnt/${cwdDrive}/`) ? "/mnt" : "";
-  }
-  return `${bashDrivePrefix}/${drive}/${rest}`;
-}
-
-function shellQuote(value: string): string {
-  return `'${value.replace(/'/g, `'\\''`)}'`;
-}
-
-function gitInit(dir: string): string {
-  const env = {
-    ...process.env,
-    GIT_AUTHOR_NAME: "t",
-    GIT_AUTHOR_EMAIL: "t@t",
-    GIT_COMMITTER_NAME: "t",
-    GIT_COMMITTER_EMAIL: "t@t",
-    GIT_CONFIG_GLOBAL: "/dev/null",
-    GIT_CONFIG_SYSTEM: "/dev/null",
-  };
-  execFileSync("git", ["init", "-b", "main", dir], { env });
-  execFileSync("git", ["-C", dir, "config", "core.autocrlf", "false"], { env });
-  execFileSync("git", ["-C", dir, "config", "core.eol", "lf"], { env });
-  writeFileSync(join(dir, "Dockerfile"), "FROM scratch\n");
-  writeFileSync(join(dir, "docker-compose.yml"), "services: {}\n");
-  writeFileSync(join(dir, "docker-compose.macos.yml"), "services: {}\n");
-  mkdirSync(join(dir, "scripts", "lib"), { recursive: true });
-  copyFileSync(join(REPO_ROOT, "scripts", "lib", "resolve-capability-compose-profiles.mjs"), join(dir, "scripts", "lib", "resolve-capability-compose-profiles.mjs"));
-  copyFileSync(join(REPO_ROOT, "scripts", "lib", "govern-capability-compose-args.mjs"), join(dir, "scripts", "lib", "govern-capability-compose-args.mjs"));
-  copyFileSync(join(REPO_ROOT, "scripts", "lib", "capability-state-hash.mjs"), join(dir, "scripts", "lib", "capability-state-hash.mjs"));
-  copyFileSync(join(REPO_ROOT, "scripts", "capability-service-catalog.generated.json"), join(dir, "scripts", "capability-service-catalog.generated.json"));
-  execFileSync("git", ["-C", dir, "add", "-A"], { env });
-  execFileSync("git", ["-C", dir, "-c", "commit.gpgsign=false", "commit", "-m", "seed"], { env });
-  return execFileSync("git", ["-C", dir, "rev-parse", "HEAD"], { env }).toString().trim();
-}
-
-/** Fake `docker` and `curl` on PATH. */
-function makeFakeBin(root: string): string {
-  const bin = join(root, "bin");
-  mkdirSync(bin, { recursive: true });
-  // docker: `build`/`up` are no-ops; any `cat /app/.dpf-source-content-hash`
-  // (the content-verify guard — step 3 `docker run` and step 7
-  // `docker compose exec`) returns a single stable hash so built==running.
-  writeFileSync(
-    join(bin, "docker"),
-    `#!/bin/sh
-env_file=
-take_env_file=0
-for arg in "$@"; do
-  if [ "$take_env_file" = 1 ]; then env_file="$arg"; take_env_file=0; continue; fi
-  if [ "$arg" = "--env-file" ]; then take_env_file=1; fi
-done
-[ -n "$DOCKER_LOG" ] && printf '%s\\n' "$*" >> "$DOCKER_LOG"
-case "$*" in
-  *".Id"*) printf "%s" "\${DPF_TEST_RELEASE_ENGINE_ID:-$DPF_TEST_RELEASE_CONFIG_DIGEST}" ;;
-  *".RepoDigests"*) printf "%s@%s\n" "$DPF_TEST_RELEASE_REPO" "\${DPF_TEST_RELEASE_REPO_ID:-$DPF_TEST_RELEASE_ENGINE_ID}" ;;
-  *".Os"*) printf "%s" "\${DPF_TEST_RELEASE_OS:-linux}" ;;
-  *".Architecture"*) printf "%s" "\${DPF_TEST_RELEASE_ARCHITECTURE:-amd64}" ;;
-  *"buildx imagetools inspect"*"--raw"*) case "$*" in *"@$DPF_TEST_RELEASE_PLATFORM_MANIFEST_DIGEST"*) printf '{"config":{"digest":"%s"}}' "\${DPF_TEST_RELEASE_REGISTRY_CONFIG_DIGEST:-$DPF_TEST_RELEASE_CONFIG_DIGEST}" ;; *) if [ "\${DPF_TEST_DUPLICATE_PLATFORM:-no}" = yes ]; then printf '{"manifests":[{"digest":"%s","platform":{"os":"%s","architecture":"%s"}},{"digest":"sha256:dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd","platform":{"os":"%s","architecture":"%s"}}]}' "$DPF_TEST_RELEASE_PLATFORM_MANIFEST_DIGEST" "$DPF_TEST_RELEASE_OS" "$DPF_TEST_RELEASE_ARCHITECTURE" "$DPF_TEST_RELEASE_OS" "$DPF_TEST_RELEASE_ARCHITECTURE"; else printf '{"manifests":[{"digest":"%s","platform":{"os":"%s","architecture":"%s"}}]}' "$DPF_TEST_RELEASE_PLATFORM_MANIFEST_DIGEST" "$DPF_TEST_RELEASE_OS" "$DPF_TEST_RELEASE_ARCHITECTURE"; fi ;; esac ;;
-  *"org.opencontainers.image.version"*) printf "%s" "$DPF_TEST_RELEASE_TAG" ;;
-  *"org.opencontainers.image.revision"*) printf "%s" "$DPF_TEST_RELEASE_SHA" ;;
-  "create "*) printf "candidate-container" ;;
-  "cp candidate-container:/dpf-release-assets/. "*)
-    for destination in "$@"; do :; done
-    # Match Docker's directory-copy semantics on both platforms.
-    mkdir -p "$destination"; cp -R "$DPF_TEST_RELEASE_ASSETS"/. "$destination/"
-    ;;
-  *"recover-human-principal-backfill-migration.mjs --verify-rolled-back"*)
-    [ "\${DPF_TEST_PRINCIPAL_VERIFY_FAIL:-no}" = "yes" ] && exit 1
-    printf "verified"
-    ;;
-  *"prisma migrate resolve --rolled-back 20260812110000_backfill_missing_human_principals"*)
-    [ "\${DPF_TEST_PRINCIPAL_RESOLVE_FAIL:-no}" = "yes" ] && exit 1
-    ;;
-  *"recover-human-principal-backfill-migration.mjs"*)
-    [ "\${DPF_TEST_PRINCIPAL_RECOVERY_DECISION:-not-needed}" = "blocked" ] && exit 1
-    if [ "\${DPF_TEST_PRINCIPAL_RECOVERY_DECISION:-not-needed}" = "recover" ]; then
-      printf "recover:11111111-1111-4111-8111-111111111111"
-    else
-      printf "not-needed"
-    fi
-    ;;
-  *"recover-inventory-snapshot-migration.mjs --verify-rolled-back"*) printf "verified" ;;
-  *"recover-inventory-snapshot-migration.mjs"*)
-    [ "\${DPF_TEST_RECOVERY_DECISION:-not-needed}" = "blocked" ] && exit 1
-    if [ "\${DPF_TEST_RECOVERY_DECISION:-not-needed}" = "recover" ]; then
-      printf "recover:11111111-1111-4111-8111-111111111111"
-    else
-      printf "not-needed"
-    fi
-    ;;
-  *"up -d --no-deps --force-recreate portal"*|*"up -d --no-deps --force-recreate sandbox"*)
-    service=
-    for service in "$@"; do :; done
-    effective_state_dir="$(sed -n 's/^DPF_STATE_DIR=//p' "$env_file" | tail -n 1)"
-    printf 'recreate service=%s DPF_STATE_DIR=%s DPF_PROMOTER_STATE_DIR=%s DPF_STATE_DIR_HOST=%s mount_source=%s\\n' \
-      "$service" "\${DPF_STATE_DIR-<unset>}" "\${DPF_PROMOTER_STATE_DIR-<unset>}" "$effective_state_dir" "$effective_state_dir" >> "$DOCKER_LOG"
-    ;;
-  *"/app/.dpf-source-content-hash"*) printf "deadbeefhash" ;;
-  "ps -a --format "*) [ -n "\${DPF_TEST_IMAGES_IN_USE:-}" ] && printf '%s\n' "$DPF_TEST_IMAGES_IN_USE" ;;
-  *"images --filter reference=ghcr.io/"*"/dpf-portal:v* --format "*) [ -n "\${DPF_TEST_PORTAL_VERSION_TAGS:-}" ] && printf '%s\n' "$DPF_TEST_PORTAL_VERSION_TAGS" ;;
-esac
-exit 0
-`,
-  );
-  // For any URL ending in /sha, report the stamped DPF_VERSION (inherited from
-  // the script's exported env) — modelling a correctly-stamped portal. Other
-  // health probes just succeed.
-  writeFileSync(
-    join(bin, "curl"),
-    '#!/bin/sh\nfor a in "$@"; do url="$a"; done\ncase "$url" in\n  */sha) printf "%s" "$DPF_VERSION" ;;\n  *) printf "ok" ;;\nesac\nexit 0\n',
-  );
-  chmodSync(join(bin, "docker"), 0o755);
-  chmodSync(join(bin, "curl"), 0o755);
-  return bin;
-}
-
-function runPromote(opts: {
-  source: string;
-  installRoot?: string;
-  backup: string;
-  targetSha: string;
-  fakeBin: string;
-  composeEnvFile?: string;
-  dockerLog?: string;
-  recoveryDecision?: "recover" | "not-needed" | "blocked";
-  principalRecoveryDecision?: "recover" | "not-needed" | "blocked";
-  principalResolveFails?: boolean;
-  principalVerifyFails?: boolean;
-  /** Newest-first `repo:tag` list the shim returns for the dpf-portal version-tag query. */
-  portalVersionTags?: string[];
-  /** `repo:tag` refs the shim reports as referenced by a container (`docker ps -a`). */
-  imagesInUse?: string[];
-  imageKeep?: number;
-  release?: {
-    tag: string; owner: string; channelDigest?: string; platformManifestDigest?: string;
-    configDigest?: string; engineImageId?: string; platformOs?: string; frozenStrata?: boolean; repoImageId?: string; registryConfigDigest?: string; duplicatePlatform?: boolean;
-    platformArchitecture?: string; enginePlatformArchitecture?: string;
-    candidateAssets: string; gitLog: string;
-  };
-}): { status: number | null; stdout: string; stderr: string } {
-  const stateDir = join(opts.backup, "state");
-  const secret = "s".repeat(32);
-  const secretPath = join(stateDir, "runtime-transition.secret");
-  writeFileSync(secretPath, secret);
-  const stateBytes = readFileSync(join(stateDir, "install-state.json"));
-  // Use the canonical migrator: a real projection hash differs from sourceHash,
-  // and the fixture must exercise that CAS binding (BI-AA6FBAD0).
-  const projection = JSON.parse(execFileSync(process.execPath, [MIGRATOR,
-    "--state", join(stateDir, "install-state.json"), "--catalog", CATALOG,
-    "--host-platform", "linux", "--host-arch", "amd64"], { encoding: "utf8" })) as { sourceHash: string; projectionHash: string };
-  const envelope = {
-    kind: "install-state-migration", version: 1, runId: "SUR-FUNCTIONAL",
-    promoterDigest: `sha256:${"d".repeat(64)}`,
-    sourceHash: projection.sourceHash,
-    projectionHash: projection.projectionHash,
-    fromSchemaVersion: 2, toSchemaVersion: 2,
-    hostIdentity: { platform: "linux", arch: "amd64", provenance: "explicit" },
-    issuedAt: new Date(Date.now() - 1_000).toISOString(),
-    expiresAt: new Date(Date.now() + 60_000).toISOString(),
-  };
-  const canonical = (value: unknown): string => Array.isArray(value)
-    ? `[${value.map(canonical).join(",")}]`
-    : value && typeof value === "object"
-      ? `{${Object.keys(value).sort().map((key) => `${JSON.stringify(key)}:${canonical((value as Record<string, unknown>)[key])}`).join(",")}}`
-      : JSON.stringify(value) ?? "null";
-  const signature = createHmac("sha256", secret).update(canonical(envelope)).digest("hex");
-  const exports = [
-    "unset DPF_STATE_DIR",
-    `export PATH=${shellQuote(toBashPath(opts.fakeBin))}:"$PATH"`,
-    `export PROMOTE_SOURCE=${shellQuote(toBashPath(opts.source))}`,
-    ...(opts.installRoot ? [`export PROMOTE_INSTALL_ROOT=${shellQuote(toBashPath(opts.installRoot))}`] : []),
-    `export PROMOTE_TARGET_SHA=${shellQuote(opts.targetSha)}`,
-    `export PROMOTE_BACKUP_PATH=${shellQuote(toBashPath(opts.backup))}`,
-    `export DPF_PROMOTER_STATE_DIR=${shellQuote(toBashPath(join(opts.backup, "state")))}`,
-    `export DPF_RUNTIME_TRANSITION_SECRET_FILE=${shellQuote(toBashPath(secretPath))}`,
-    "export DPF_SELF_UPGRADE_RUN_ID='SUR-FUNCTIONAL'",
-    `export DPF_PROMOTER_DIGEST=${shellQuote(envelope.promoterDigest)}`,
-    `export DPF_INSTALL_STATE_MIGRATION_ENVELOPE=${shellQuote(Buffer.from(JSON.stringify(envelope)).toString("base64url"))}`,
-    `export DPF_INSTALL_STATE_MIGRATION_SIGNATURE=${shellQuote(signature)}`,
-    "export PROMOTE_HEALTH_URL='http://127.0.0.1:9/api/health'",
-    "export PROMOTE_COMPOSE_PROJECT='dpf-functest'",
-    ...(opts.recoveryDecision
-      ? [`export DPF_TEST_RECOVERY_DECISION=${shellQuote(opts.recoveryDecision)}`]
-      : []),
-    ...(opts.principalRecoveryDecision
-      ? [`export DPF_TEST_PRINCIPAL_RECOVERY_DECISION=${shellQuote(opts.principalRecoveryDecision)}`]
-      : []),
-    ...(opts.principalResolveFails ? ["export DPF_TEST_PRINCIPAL_RESOLVE_FAIL=yes"] : []),
-    ...(opts.principalVerifyFails ? ["export DPF_TEST_PRINCIPAL_VERIFY_FAIL=yes"] : []),
-    ...(opts.composeEnvFile
-      ? [`export PROMOTE_COMPOSE_ENV_FILE=${shellQuote(toBashPath(opts.composeEnvFile))}`]
-      : []),
-    ...(opts.dockerLog ? [`export DOCKER_LOG=${shellQuote(toBashPath(opts.dockerLog))}`] : []),
-    ...(opts.portalVersionTags ? [`export DPF_TEST_PORTAL_VERSION_TAGS=${shellQuote(opts.portalVersionTags.join("\n"))}`] : []),
-    ...(opts.imagesInUse ? [`export DPF_TEST_IMAGES_IN_USE=${shellQuote(opts.imagesInUse.join("\n"))}`] : []),
-    ...(opts.imageKeep !== undefined ? [`export PROMOTE_IMAGE_KEEP=${opts.imageKeep}`] : []),
-    ...(opts.release ? [
-      "export DPF_PROMOTION_MODE=release",
-      `export DPF_RELEASE_TAG=${shellQuote(opts.release.tag)}`,
-      ...(opts.release.configDigest
-        ? [`export DPF_RELEASE_CONFIG_DIGEST=${shellQuote(opts.release.configDigest)}`]
-        : ["unset DPF_RELEASE_CONFIG_DIGEST"]),
-      ...(opts.release.frozenStrata === false || !opts.release.configDigest ? ["unset DPF_RELEASE_CHANNEL_DIGEST DPF_RELEASE_PLATFORM_MANIFEST_DIGEST DPF_RELEASE_PLATFORM_OS DPF_RELEASE_PLATFORM_ARCHITECTURE"] : [`export DPF_RELEASE_CHANNEL_DIGEST=${shellQuote(opts.release.channelDigest ?? "")}`, `export DPF_RELEASE_PLATFORM_MANIFEST_DIGEST=${shellQuote(opts.release.platformManifestDigest ?? "")}`, `export DPF_RELEASE_PLATFORM_OS=${shellQuote(opts.release.platformOs ?? "linux")}`, `export DPF_RELEASE_PLATFORM_ARCHITECTURE=${shellQuote(opts.release.platformArchitecture ?? "amd64")}`]),
-      `export GHCR_OWNER=${shellQuote(opts.release.owner)}`,
-      `export DPF_TEST_RELEASE_SHA=${shellQuote(opts.targetSha)}`,
-      `export DPF_TEST_RELEASE_TAG=${shellQuote(opts.release.tag)}`,
-      `export DPF_TEST_RELEASE_CONFIG_DIGEST=${shellQuote(opts.release.configDigest ?? `sha256:${"c".repeat(64)}`)}`,
-      `export DPF_TEST_RELEASE_ENGINE_ID=${shellQuote(opts.release.engineImageId ?? opts.release.configDigest ?? `sha256:${"c".repeat(64)}`)}`,
-      `export DPF_TEST_RELEASE_REPO_ID=${shellQuote(opts.release.repoImageId ?? "")}`, `export DPF_TEST_RELEASE_REGISTRY_CONFIG_DIGEST=${shellQuote(opts.release.registryConfigDigest ?? "")}`,
-      `export DPF_TEST_DUPLICATE_PLATFORM=${opts.release.duplicatePlatform ? "yes" : "no"}`,
-      `export DPF_TEST_RELEASE_REPO=ghcr.io/${shellQuote(opts.release.owner)}/dpf-portal`,
-      `export DPF_TEST_RELEASE_PLATFORM_MANIFEST_DIGEST=${shellQuote(opts.release.platformManifestDigest ?? `sha256:${"b".repeat(64)}`)}`,
-      `export DPF_TEST_RELEASE_OS=${shellQuote(opts.release.platformOs ?? "linux")}`,
-      `export DPF_TEST_RELEASE_ARCHITECTURE=${shellQuote(opts.release.enginePlatformArchitecture ?? opts.release.platformArchitecture ?? "amd64")}`,
-      `export DPF_TEST_RELEASE_ASSETS=${shellQuote(toBashPath(opts.release.candidateAssets))}`,
-      `export GIT_LOG=${shellQuote(toBashPath(opts.release.gitLog))}`,
-      "export PROMOTE_COMPOSE_FILES='docker-compose.yml docker-compose.release.yml'",
-    ] : []),
-  ];
-  const r = spawnSync(BASH_COMMAND, ["-lc", `${exports.join("\n")}\nexec bash ${shellQuote(toBashPath(SCRIPT))} --self-upgrade`], {
-    encoding: "utf8",
-  });
-  return { status: r.status, stdout: r.stdout ?? "", stderr: r.stderr ?? "" };
-}
-
-/** Scratch root with a clean git source dir (src/) + isolated bin/ and backup/. */
-function makeScratch(): { root: string; source: string; backup: string; fakeBin: string; head: string } {
-  // macOS exposes /var as a symlink to /private/var. Use the canonical parent
-  // so the production state-path escape guard sees the same path before and
-  // after realpath resolution.
-  const root = mkdtempSync(join(realpathSync(tmpdir()), "dpf-promote-"));
-  const source = join(root, "src");
-  mkdirSync(source, { recursive: true });
-  const head = gitInit(source);
-  const fakeBin = makeFakeBin(root);
-  const backup = join(root, "backup");
-  const stateDir = join(backup, "state");
-  mkdirSync(stateDir, { recursive: true });
-  // Produce state through the canonical legacy migrator so capability closure
-  // cannot drift from the runtime resolver.
-  writeFileSync(join(stateDir, "install-state.json"), JSON.stringify({
-    schemaVersion: 1, installerVersion: "functional", platform: "linux", arch: "amd64",
-    installPath: "/opt/dpf", stateDir: "/dpf-state", composeProjectName: "dpf",
-  }));
-  execFileSync(process.execPath, [MIGRATOR, "--state", join(stateDir, "install-state.json"),
-    "--catalog", CATALOG, "--host-platform", "linux", "--host-arch", "amd64", "--write"], { encoding: "utf8" });
-  return { root, source, backup, fakeBin, head };
-}
+import { BASH_OK, GIT_OK, PROMOTE_TEST_TIMEOUT_MS, makeScratch, runPromote, toBashPath } from "./promote-script-functional.test-support";
 
 describe.skipIf(!BASH_OK || !GIT_OK)("promote.sh — real-script functional run", () => {
   it.each([
@@ -395,69 +125,6 @@ describe.skipIf(!BASH_OK || !GIT_OK)("promote.sh — real-script functional run"
       // sha-verify completed against the SHA actually built (HEAD).
       expect(r.stdout).toContain(`step=done target=${head}`);
       expect(r.stderr).not.toContain("does not match promote target");
-    } finally {
-      rmSync(root, { recursive: true, force: true });
-    }
-  }, PROMOTE_TEST_TIMEOUT_MS);
-
-  it("sweeps dangling images and build cache after a successful promote", () => {
-    const { root, source, backup, fakeBin, head } = makeScratch();
-    const dockerLog = join(root, "docker.log");
-    try {
-      const r = runPromote({ source, backup, targetSha: head, fakeBin, dockerLog });
-      expect(r.status).toBe(0);
-      expect(r.stdout).toContain("step=cleanup");
-      const dockerCalls = readFileSync(dockerLog, "utf8");
-      // Best-effort disk reclaim on the success path.
-      // Dangling images: removed outright (superseded previous portal version).
-      expect(dockerCalls).toContain("image prune -f");
-      // Build cache: BOUNDED, not wiped — kept as a rebuild-speed asset under a cap.
-      expect(dockerCalls).toContain("builder prune -f --keep-storage");
-      // Conservative scope: never the destructive `-a` (would delete in-use tagged images).
-      expect(dockerCalls).not.toContain("image prune -a");
-      // Volumes are operator state — cleanup must never touch them.
-      expect(dockerCalls).not.toContain("volume");
-      // BI-9B7FC928: also reclaims the TAGGED ephemeral verify/compare/build-test
-      // images that dangling-prune misses (the ~3.7 GB/upgrade leak). It queries
-      // by exact ephemeral naming so the running dpf-portal image is never hit.
-      expect(dockerCalls).toContain("images --filter reference=dpf-*-build-test");
-      expect(dockerCalls).toContain("images --filter reference=dpf-*-build-compare");
-      expect(dockerCalls).toContain("images --filter reference=dpf-*:verify");
-    } finally {
-      rmSync(root, { recursive: true, force: true });
-    }
-  }, PROMOTE_TEST_TIMEOUT_MS);
-
-  it("bounds superseded version tags per self-upgrade image: keeps in-use + PROMOTE_IMAGE_KEEP newest, removes the rest by tag", () => {
-    // Every upgrade tags a fresh portal/postgres/promoter/sandbox and the tag
-    // keeps it out of dangling-prune, so a dev install piled up 92 portal tags
-    // (388 GB reclaimable) in 16 days. The shim answers the portal query
-    // newest-first; the running tag sits in the middle to prove in-use wins
-    // over recency and never counts against the keep budget.
-    const { root, source, backup, fakeBin, head } = makeScratch();
-    const dockerLog = join(root, "docker.log");
-    const repo = "ghcr.io/opendigitalproductfactory/dpf-portal";
-    const tags = ["v2026.09.06-e.1", "v2026.09.06-d.1", "v2026.09.05-running.1", "v2026.09.04-c.1", "v2026.09.03-b.1", "v2026.09.02-a.1"]
-      .map((tag) => `${repo}:${tag}`);
-    try {
-      const r = runPromote({
-        source, backup, targetSha: head, fakeBin, dockerLog,
-        portalVersionTags: tags,
-        imagesInUse: [`${repo}:v2026.09.05-running.1`, "ghcr.io/opendigitalproductfactory/dpf-sandbox:v2026.09.05-running.1"],
-        imageKeep: 2,
-      });
-      expect(r.status).toBe(0);
-      const dockerCalls = readFileSync(dockerLog, "utf8").split("\n");
-      const removed = dockerCalls.filter((line) => line.startsWith("rmi ") && line.includes("dpf-portal:v"));
-      // Newest two survive as rollback margin, the in-use one survives regardless of age.
-      expect(removed).toEqual([`rmi ${repo}:v2026.09.04-c.1`, `rmi ${repo}:v2026.09.03-b.1`, `rmi ${repo}:v2026.09.02-a.1`]);
-      expect(dockerCalls).not.toContain(`rmi ${repo}:v2026.09.05-running.1`);
-      // Every self-upgrade repo is swept, and by version-tag reference only.
-      for (const name of ["dpf-portal", "dpf-postgres", "dpf-promoter", "dpf-sandbox"]) {
-        expect(dockerCalls).toContain(`images --filter reference=ghcr.io/opendigitalproductfactory/${name}:v* --format {{.Repository}}:{{.Tag}}`);
-      }
-      // Untag by reference, never force-remove by id: a layer shared with a kept tag must survive.
-      expect(removed.every((line) => !line.includes(" -f "))).toBe(true);
     } finally {
       rmSync(root, { recursive: true, force: true });
     }
